Avoid Throwable gson serialization with transient
Trying to serialize a java.lang.Throwable will raise a warning about
illegal reflection. Marking the Throwable being serialized as `transient`
will cause gson to ignore this field during serialization, it's not part
of the notice payload anyway.
https://github.com/google/gson/blob/master/UserGuide.md#java-modifier-exclusion

Note on transient keyword:
> `transient` is a Java keyword which marks a member variable not to be
serialized when it is persisted to streams of bytes. When an object is
transferred through the network, the object needs to be 'serialized'.
source: [wikibooks - Java Programming](https://en.wikibooks.org/wiki/Java_Programming/Keywords/transient)